move api lookups to utils
