+++ packages/destination-actions/src/destinations/engage/sendgrid/__tests__/send-email.test.ts
@@ -5,7 +5,7 @@ import Sendgrid from '..'
 import { FLAGON_NAME_LOG_ERROR, FLAGON_NAME_LOG_INFO, SendabilityStatus } from '../../utils'
 import { loggerMock, expectErrorLogged, expectInfoLogged } from '../../utils/testUtils'
 import { insertEmailPreviewText } from '../sendEmail/insertEmailPreviewText'
-import { FLAGON_NAME_DATA_FEEDS } from '../previewApiLookup'
+import { FLAGON_NAME_DATA_FEEDS } from '../../utils/apiLookups'
 
 const sendgrid = createTestIntegration(Sendgrid)
 const timestamp = new Date().toISOString()
