+++ packages/destination-actions/src/destinations/engage/sendgrid/previewApiLookup/index.ts
@@ -1,2 +1,1 @@
 export * from './actionDefinition'
-export * from './api-lookups'
