+++ packages/destination-actions/src/destinations/engage/sendgrid/sendEmail/actionDefinition.ts
@@ -1,7 +1,7 @@
 import { ActionDefinition } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
-import { apiLookupActionFields } from '../previewApiLookup'
+import { apiLookupActionFields } from '../../utils/apiLookups'
 import { SendEmailPerformer } from './SendEmailPerformer'
 
 export const actionDefinition: ActionDefinition<Settings, Payload> = {
